requirements.txt
pandas
numpy
matplotlib
seaborn
tldextract
aiohttp
scikit-learn
mlflow
docker
shap
xgboost
uvicorn[standard]
pymongo
ipykernel
hyperopt
evidently
azure-storage-blob